requirements.txt
aiosmtplib<=2.0.2
aioredis<=2.0.1
alembic<=1.13.2
annotated-types<=0.7.0
anyio<=4.6.0
asyncpg<=0.30.0
blinker<=1.8.2
bcrypt<=4.2.0
certifi<=2024.8.30
cffi<=2.0.0
charset-normalizer<=3.3.2
celery<=5.4.0
click<=8.1.7
colorama<=0.4.6
cryptography<=46.0.2
dnspython<=2.6.1
ecdsa<=0.19.0
email_validator<=2.2.0
fastapi<=0.115.0
fastapi-cli<=0.0.5
fastapi-mail<=1.4.1
firebase-admin<=6.6.0    
geoalchemy2<=0.15.2
greenlet<=3.1.1
h11<=0.14.0
httpcore<=1.0.5
httptools<=0.7.1
httpx<=0.27.2
idna<=3.10
itsdangerous<=2.2.0
Jinja2<=3.1.4
mailjet-rest<=1.3.4
Mako<=1.3.5
markdown-it-py<=3.0.0
MarkupSafe<=2.1.5
mdurl<=0.1.2
numpy<=2.1.2
orjson<=3.10.7
pandas<=2.2.3
passlib[bcrypt]<=1.7.4
paystackapi<=2.1.0
psycopg2-binary<=2.9.10
pyasn1<=0.6.1
pycparser<=2.22
pydantic[email]<=2.9.2
pydantic-extra-types<=2.9.0
pydantic-settings<=2.5.2
pydantic_core<=2.23.4
Pygments<=2.18.0
pyotp<=2.9.0
python-dateutil<=2.9.0.post0
python-dotenv<=1.0.1
python-jose[cryptography]<=3.3.0
python-multipart<=0.0.12
pytz<=2024.2
PyYAML<=6.0.2
redis<=5.1.1
requests<=2.32.3
rich<=13.8.1
rsa<=4.9
shellingham<=1.5.4
six<=1.16.0
sniffio<=1.3.1
SQLAlchemy<=2.0.35
SQLAlchemy-Utils<=0.41.2
starlette<=0.38.5
typer<=0.12.5
typing_extensions<=4.12.2
tzdata<=2024.2
ujson<=5.10.0
urllib3<=2.2.3
uvicorn<=0.30.6
watchfiles<=0.24.0
websockets<=13.0.1